+++ EssentialFeed/EssentialFeediOSTests/FeedUI/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,39 @@
+//
+//  FeedImageCell+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Radosław Serek on 05/02/2023.
+//
+
+import EssentialFeediOS
+import XCTest
+
+extension FeedImageCell {
+    var isShowingLocation: Bool {
+        return !locationContainer.isHidden
+    }
+    
+    var locationText: String? {
+        return locationLabel.text
+    }
+    
+    var descriptionText: String? {
+        return descriptionLabel.text
+    }
+    
+    var isShowingImageLoadingIndicator: Bool {
+        return feedImageContainer.isShimmering
+    }
+    
+    var renderedImage: Data? {
+        return feedImageView.image?.pngData()
+    }
+    
+    var isShowingRetryButton: Bool {
+        return !feedImageRetryButton.isHidden
+    }
+    
+    func simulateRetryAction() {
+        feedImageRetryButton.simulateTap()
+    }
+}
